+++ src/pages/FriendsPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
 
-export default function FollowersPage() {
+export default function FriendsPage() {
   return (
     <div>
       <h1>팔로우 목록</h1>
