--- src/pages/NotificationsPage.js
@@ -1,11 +1,0 @@
-import React from "react";
-import {Link} from "react-router-dom";
-
-export default function NotificationsPage() {
-    return (
-        <div>
-            <h1>알림 페이지</h1>
-            <Link to="/">뒤로가기</Link>
-        </div>
-    )
-}
