--- src/pages/SearchEventPage.js
@@ -1,11 +1,0 @@
-import React from "react";
-import {Link} from "react-router-dom";
-
-export default function SearchEventPage() {
-    return (
-        <div>
-            <h1>이벤트 검색</h1>
-            <Link to="/">뒤로가기</Link>
-        </div>
-    )
-}
