+++ src/pages/SearchFriendsPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
 
-export default function SearchFollowerPage() {
+export default function SearchFriendsPage() {
   return (
     <div>
       <h1>팔로우 검색/추가 페이지</h1>
